Added envdte dependencies: to GX18u7. It seems envdte is not registered in GAC since VS 2017

+++ LsiGxUtilidades-Gx16/Utilidades/VS/VisualStudio.cs
@@ -152,7 +152,7 @@ namespace LSI.Packages.Extensiones.Utilidades.VS
             {
                 Dte.MainWindow.WindowState = vsWindowState.vsWindowStateMaximize;
                 Dte.MainWindow.Activate();
-                SetForegroundWindow(Dte.MainWindow.HWnd);
+                SetForegroundWindow((int)Dte.MainWindow.HWnd);
             }
             catch {}
         }
